+++ SOTFEdit/Model/SaveData/NamedIntData.cs
@@ -1,0 +1,6 @@
+﻿using System.Diagnostics.CodeAnalysis;
+
+namespace SOTFEdit.Model.SaveData;
+
+[SuppressMessage("ReSharper", "NotAccessedPositionalProperty.Global")]
+public record NamedIntData(string SaveObjectNameId, int SaveValue);